Finished the homepage and made the section into components for readability and maintainability
