+++ src/components/pages/Home/Works/Works.jsx
@@ -1,0 +1,23 @@
+import { NavLink } from "react-router-dom";
+import { Button } from "../../../Button";
+
+function Works() {
+  return (
+    <section className="relative bg-featured-works-pattern bg-center w-full h-full p-16 bg-cover flex flex-col items-center">
+      <div className="absolute inset-0 bg-rgbaLightRed z-10"></div>
+      <div className="text-2xl lg:text-5xl text-white uppercase drop-shadow-neonWhite w-full h-96 md:h-36  max-w-[120rem] font-daysOne text-left mt-1 lg:mt-2 z-20 relative">
+        <h1>Check out our featured works!</h1>
+        <p className="text-white text-xl drop-shadow-none font-lexendDeca z-20 lowercase">
+          &quot;Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do{" "}
+          <br /> eiusmod tempor incididunt ut labore et dolore magna
+          aliqua.&quot;
+        </p>
+      </div>
+      <Button className="bg-transparent uppercase shadow-glowWhite border-4 text-3xl text-white mt-4 z-20 self-end h-[70px] w-[300px]">
+        <NavLink to="/gallery">Visit Gallery</NavLink>
+      </Button>
+    </section>
+  );
+}
+
+export default Works;
